+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 - Updated vuejs webpack template from 1.2.7 to 1.2.8
 - Removed CSS from base app component
+- Removed demo components from `HelloWorld`and updated tests
 ### Deprecated
 ### Removed
 ### Fixed